moved UploadDataFromPath into its own file
